ts7400v2: Fixup GPIOs with no internal pull
Needed for the move to kernel 5.10. The issue is the GPIO used
to drive the some signals do not have an internal pull, keeper
only. This means for reliable USB operation, U-Boot needs to set
this pin as an output and drive it before the kernel starts. The
same is done for keeping CAN transceiver disabled.

This also gives a proper power on and reset to ethernet to ensure
that if U-Boot needs ethernet that is really is completely valid.

Signed-off-by: Kris Bahnsen <kris@embeddedTS.com>